function_app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob

#Main.py
tqdm

#Judge.py
openai
python-dotenv

#mp4_2_text.py
azure-cognitiveservices-speech

#Text_extractor
PyMuPDF
